Modification of ignbin.c (compatibility windows).

+++ scilab/modules/randlib/src/c/ignbin.c
@@ -141,7 +141,7 @@ S20:
     nsave = *n;
     if(xnp < 30.0) goto S140;
     ffm = xnp+p;
-    m = ffm;
+    m = (int)ffm;
     fm = m;
     xnpq = xnp*q;
     p1 = (int) (2.195*sqrt(xnpq)-4.6*q)+0.5;
@@ -166,7 +166,7 @@ S30:
      TRIANGULAR REGION
 */
     if(u > p1) goto S40;
-    ix = xm-p1*v+u;
+    ix = (int)(xm-p1*v+u);
     goto S170;
 S40:
 /*
@@ -176,14 +176,14 @@ S40:
     x = xl+(u-p1)/c;
     v = v*c+1.0-Abs(xm-x)/p1;
     if(v > 1.0 || v <= 0.0) goto S30;
-    ix = x;
+    ix = (int)x;
     goto S70;
 S50:
 /*
      LEFT TAIL
 */
     if(u > p3) goto S60;
-    ix = xl+log(v)/xll;
+    ix = (int)(xl+log(v)/xll);
     if(ix < 0) goto S30;
     v *= ((u-p2)*xll);
     goto S70;
@@ -191,7 +191,7 @@ S60:
 /*
      RIGHT TAIL
 */
-    ix = xr-log(v)/xlr;
+    ix = (int)(xr-log(v)/xlr);
     if(ix > *n) goto S30;
     v *= ((u-p3)*xlr);
 S70:
